pg-data-files - Move both dictionary and word tree data file managers to the same factory (#11, #13).
