+++ pg-data-files/DataFiles/DictionariesDataFactory.cs
@@ -1,6 +1,8 @@
-﻿using System.Text;
+﻿using PG.Data.Files.DataFiles.Dictionaries;
+using PG.Data.Files.DataFiles.WordTrees;
+using System.Text;
 
-namespace PG.Data.Files.Dictionaries
+namespace PG.Data.Files.DataFiles
 {
 	/// <summary>
 	/// Factory for creating instances of <see cref="IDictionariesData"/>.
@@ -11,14 +13,24 @@ namespace PG.Data.Files.Dictionaries
 		/// Creates an instance of <see cref="IDictionariesData"/> for the specified file depending on its extension.
 		/// </summary>
 		/// <exception cref="NotSupportedException">Thrown when the file extension is not supported.</exception>
-		public IDictionariesData CreateForFile(string filePath, Encoding encoding)
+		public IDictionariesData CreateForDictionaryFile(string filePath, Encoding encoding)
 		{
 			string extension = Path.GetExtension(filePath);
 
 			return extension.ToLower() switch
 			{
 				".txt" => new TextDictionaryFile(filePath, encoding),
-				//".csv" => new CsvDictionariesData(filePath, encoding),
+				_ => throw new NotSupportedException($"File extension '{extension}' is not supported.")
+			};
+		}
+
+		public IWordTreeData CreateForWordTreeFile(string filePath)
+		{
+			string extension = Path.GetExtension(filePath);
+
+			return extension.ToLower() switch
+			{
+				".gz" => new BinaryWordTreeFile(filePath),
 				_ => throw new NotSupportedException($"File extension '{extension}' is not supported.")
 			};
 		}
