+++ pg-data-files/DataFiles/WordTrees/IWordTreeData.cs
@@ -1,0 +1,11 @@
+﻿using PG.Entities.WordTrees;
+
+namespace PG.Data.Files.DataFiles.WordTrees
+{
+    public interface IWordTreeData
+    {
+        WordDictionaryTree FetchTree();
+
+        void SaveTree(WordDictionaryTree tree);
+    }
+}
