--- pg-data-files/Dictionaries/IDictionariesData.cs
@@ -1,10 +1,0 @@
-﻿namespace PG.Data.Files.Dictionaries
-{
-	/// <summary>
-	/// Represents a data source for dictionaries to get words from.
-	/// </summary>
-	public interface IDictionariesData
-	{
-		IEnumerable<string> FetchAllWords();
-	}
-}
