--- pg-data-files/WordTrees/IWordTreeData.cs
@@ -1,11 +1,0 @@
-﻿using PG.Entities.WordTrees;
-
-namespace PG.Data.Files.WordTrees
-{
-	public interface IWordTreeData
-	{
-		WordDictionaryTree FetchTree();
-
-		void SaveTree(WordDictionaryTree tree);
-	}
-}
