+++ pg-logic/Passwords/Loader/IDictionaryLoaderFactory.cs
@@ -1,9 +1,11 @@
-﻿using System.Text;
+﻿using PG.Data.Files.DataFiles.WordTrees;
+using System.Text;
 
 namespace PG.Logic.Passwords.Loader
 {
 	public interface IDictionaryLoaderFactory
 	{
-		IDictionaryLoader Create(string filePath, Encoding encoding);
+		IDictionaryLoader CreateForDictionary(string filePath, Encoding encoding);
+		IWordTreeData CreateForWordTree(string filePath);
 	}
 }
