+++ pg-logic/Passwords/Loader/WordDictionaryLoader.cs
@@ -1,11 +1,11 @@
-﻿using PG.Data.Files.Dictionaries;
+﻿using PG.Data.Files.DataFiles.Dictionaries;
 using PG.Entities.WordTrees;
 using PG.Shared.Extensions;
 using static PG.Logic.ErrorHandling.BusinessExceptions;
 
 namespace PG.Logic.Passwords.Loader
 {
-	public class WordDictionaryLoader(IDictionariesData data) : IDictionaryLoader
+    public class WordDictionaryLoader(IDictionariesData data) : IDictionaryLoader
 	{
 		private const int MINIMUM_WORD_LENGTH = 2;
 		private readonly HashSet<char> VOWEL_AND_DIACRITIC_CHARS = [
