+++ pg-tests/Business/Passwords/Generators/Mockups/DictionaryDataMockup.cs
@@ -1,8 +1,8 @@
-﻿using PG.Data.Files.Dictionaries;
+﻿using PG.Data.Files.DataFiles.Dictionaries;
 
 namespace PG.Tests.Business.Passwords.Generators.Mockups
 {
-	internal class DictionaryDataMockup(string[] words) : IDictionariesData
+    internal class DictionaryDataMockup(string[] words) : IDictionariesData
 	{
 		private readonly string[] _words = words;
